# openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "OpenAPI Specification",
    "description": "OpenAPI Documentation for Food Planner Application",
    "contact": {
      "name": "Gerzsenyi Patrik",
      "email": "q71lux@inf.elte.hu"
    },
    "version": "0.0.1-SNAPSHOT"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
  "paths": {
    "/shoppinglist": {
      "get": {
        "tags": [
          "Shopping List"
        ],
        "summary": "Retrieves all shopping list records",
        "operationId": "getAllShoppingLists",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ShoppingList"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Shopping List"
        ],
        "summary": "Creates a new shopping list record",
        "operationId": "addShoppingList",
        "parameters": [
          {
            "name": "shoppingListDTO",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/ShoppingListDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ShoppingList"
                }
              }
            }
          }
        }
      }
    },
    "/recipe": {
      "get": {
        "tags": [
          "Recipe"
        ],
        "summary": "Retrieves all recipes",
        "operationId": "getAllRecipes",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Recipe"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Recipe"
        ],
        "summary": "Creates a new recipe",
        "operationId": "addRecipe",
        "parameters": [
          {
            "name": "recipeDTO",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/RecipeDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Recipe"
                }
              }
            }
          }
        }
      }
    },
    "/pantry": {
      "get": {
        "tags": [
          "Pantry"
        ],
        "summary": "Retrieves all pantry records",
        "operationId": "getAllPantries",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Pantry"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Pantry"
        ],
        "summary": "Creates a new pantry record",
        "operationId": "addPantry",
        "parameters": [
          {
            "name": "pantryDTO",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/PantryDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Pantry"
                }
              }
            }
          }
        }
      }
    },
    "/mealplan": {
      "get": {
        "tags": [
          "Meal Plan"
        ],
        "summary": "Retrieves a list of all meal plans",
        "operationId": "getAllMealPlans",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MealPlan"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Meal Plan"
        ],
        "summary": "Creates a new meal plan",
        "operationId": "addMealPlan",
        "parameters": [
          {
            "name": "mealPlanDTO",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/MealPlanDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/MealPlan"
                }
              }
            }
          }
        }
      }
    },
    "/shoppinglist/{id}": {
      "get": {
        "tags": [
          "Shopping List"
        ],
        "summary": "Retrieves a shopping list record by id",
        "operationId": "getShoppingList",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ShoppingList"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Shopping List"
        ],
        "summary": "Deletes a shopping list record by id",
        "operationId": "deleteShoppingList",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "patch": {
        "tags": [
          "Shopping List"
        ],
        "summary": "Updates a shopping list record by id",
        "operationId": "updateShoppingList",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "shoppingListDTO",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/ShoppingListDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ShoppingList"
                }
              }
            }
          }
        }
      }
    },
    "/recipe/{id}": {
      "get": {
        "tags": [
          "Recipe"
        ],
        "summary": "Retrieves a recipe by id",
        "operationId": "getRecipe",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Recipe"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Recipe"
        ],
        "summary": "Deletes a recipe by id",
        "operationId": "deleteRecipe",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "patch": {
        "tags": [
          "Recipe"
        ],
        "summary": "Updates a recipe by id",
        "operationId": "updateRecipe",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "recipeDTO",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/RecipeDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Recipe"
                }
              }
            }
          }
        }
      }
    },
    "/pantry/{id}": {
      "get": {
        "tags": [
          "Pantry"
        ],
        "summary": "Retrieves a pantry by id",
        "operationId": "getPantry",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Pantry"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Pantry"
        ],
        "summary": "Deletes a pantry by id",
        "operationId": "deletePantry",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "patch": {
        "tags": [
          "Pantry"
        ],
        "summary": "Updates a pantry by id",
        "operationId": "updatePantry",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "pantryDTO",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/PantryDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Pantry"
                }
              }
            }
          }
        }
      }
    },
    "/mealplan/{id}": {
      "get": {
        "tags": [
          "Meal Plan"
        ],
        "summary": "Retrieves a meal plan by id",
        "operationId": "getMealPlan",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/MealPlan"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Meal Plan"
        ],
        "summary": "Deletes a meal plan by id",
        "operationId": "deleteMealPlan",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "patch": {
        "tags": [
          "Meal Plan"
        ],
        "summary": "Updates a meal plan by id",
        "operationId": "updateMealPlan",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "mealPlanDTO",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/MealPlanDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/MealPlan"
                }
              }
            }
          }
        }
      }
    },
    "/kitchen": {
      "get": {
        "tags": [
          "Kitchen"
        ],
        "summary": "Retrieves a list of all kitchen",
        "operationId": "getAllKitchen",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Kitchen"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/ingredient": {
      "get": {
        "tags": [
          "Ingredient"
        ],
        "summary": "Retrieves a list of all ingredient ingredients",
        "operationId": "getAllIngredients",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Ingredient"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/ingredient/category": {
      "get": {
        "tags": [
          "Ingredient Category"
        ],
        "summary": "Retrieves a list of all ingredient categories",
        "operationId": "getAllIngredientCategories",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/IngredientCategory"
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ShoppingListDTO": {
        "type": "object",
        "properties": {
          "ingredientId": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
      "Ingredient": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "category": {
            "$ref": "#/components/schemas/IngredientCategory"
          },
          "recipes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Recipe"
            },
            "uniqueItems": true
          }
        }
      },
      "IngredientCategory": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "ingredients": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Ingredient"
            },
            "uniqueItems": true
          }
        }
      },
      "Kitchen": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "recipes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Recipe"
            },
            "uniqueItems": true
          }
        }
      },
      "MealPlan": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "recipe": {
            "$ref": "#/components/schemas/Recipe"
          },
          "startDate": {
            "type": "string",
            "format": "date-time"
          },
          "endDate": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "Recipe": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "kitchen": {
            "$ref": "#/components/schemas/Kitchen"
          },
          "ingredients": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Ingredient"
            },
            "uniqueItems": true
          },
          "mealPlans": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MealPlan"
            },
            "uniqueItems": true
          }
        }
      },
      "ShoppingList": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "ingredient": {
            "$ref": "#/components/schemas/Ingredient"
          }
        }
      },
      "RecipeDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "kitchenId": {
            "type": "string",
            "format": "uuid"
          },
          "ingredientIds": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "uniqueItems": true
          }
        }
      },
      "PantryDTO": {
        "type": "object",
        "properties": {
          "ingredientId": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
      "Pantry": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "ingredient": {
            "$ref": "#/components/schemas/Ingredient"
          }
        }
      },
      "MealPlanDTO": {
        "type": "object",
        "properties": {
          "recipeId": {
            "type": "string",
            "format": "uuid"
          },
          "startDate": {
            "type": "string",
            "format": "date-time"
          },
          "endDate": {
            "type": "string",
            "format": "date-time"
          }
        }
      }
    }
  }
}